+++ partgraph_c/PrimMaker.h
@@ -13,10 +13,10 @@ class PrimMaker
 {
 public:
 	static std::shared_ptr<TopoShape> Plane(double x, double y, double z, double nx, double ny, double nz,
-		uint16_t op_id = 0, const std::shared_ptr<breptopo::TopoNaming>& tn = nullptr);
+		uint32_t op_id = 0, const std::shared_ptr<breptopo::TopoNaming>& tn = nullptr);
 
 	static std::shared_ptr<TopoShape> Box(double dx, double dy, double dz, 
-		uint16_t op_id = 0, const std::shared_ptr<breptopo::TopoNaming>& tn = nullptr);
+		uint32_t op_id = 0, const std::shared_ptr<breptopo::TopoNaming>& tn = nullptr);
 	static std::shared_ptr<TopoShape> Cylinder(double radius, double length);
 	static std::shared_ptr<TopoShape> Cone(double r1, double r2, double height);
 	static std::shared_ptr<TopoShape> Sphere(double radius);
